+++ msvc-productos/src/main/java/com/minsait/msvc/service/ProductoService.java
@@ -2,6 +2,7 @@ package com.minsait.msvc.service;
 
 import com.minsait.msvc.models.entity.Producto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface ProductoService {
@@ -9,7 +10,7 @@ public interface ProductoService {
     Producto findById(Long id);
     Producto save(Producto producto);
     boolean deleteById(Long id);
-    double checkPrice(Long id);
+    BigDecimal checkPrice(Long id);
     //TODO: Realizar metodo en productoRepository para buscar por el nombre del producto
     //Producto findByName(String nombre);
 }
