+++ msvc-productos/src/main/java/com/minsait/msvc/service/ProductoServiceImpl.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -28,7 +29,7 @@ public class ProductoServiceImpl implements ProductoService{
 
     @Override
     @Transactional(readOnly = true)
-    public double checkPrice(Long id) {
+    public BigDecimal checkPrice(Long id) {
         return productoRepository.findById(id).get().getPrecio();
     }
